Config File Loading, Repackaging Combat source folder, Removed external junk files
Added a Config loader, data can be read off of a flat file which can also be manipulated on the live game on demand.

Fixed the packaging in the Combat source folder to represent a more accurate pack naming structure.

Also removed a few extra file that served no purpose as well.