Primitive Object changes, Junk Removing
Junk removing is going to be a work in progress, as we need to understand what exists in this revision and determine its value and if and when we need to rewrite the content accordingly.

Primitive objects has been updated to more suitable objects. It allows us to write Java a little bit more effeciently, etc..