Cleanup, Optimizing, Changes
Lot of Cleanup, and optimizing.

Got rid of the World Thread class and introduced a more simplified system in the World class for it's relations.

Clean, optimizing, code moving, etc...

Going to rework the World Player index thing as seen in PublicChatMessage. There's quite a few so the rest will be sent in the next update.