Skill class improvements & Doc
Adds some tweaks to the Skill class & documentation as well.
Should help new people understand the functions better now.

The xp counter will now have 1,434 xp when a new account is made. This is cause we have experience in Hitpoints and Herblore, so only makes sense to actually add this for new players.